Exercise 10?

+++ src/main/java/io/codelex/arithmetic/practice/Geometry.java
@@ -5,18 +5,18 @@ package io.codelex.arithmetic.practice;
 import java.math.BigDecimal;
 
 class Geometry {
-    static BigDecimal areaOfCircle(BigDecimal radius) {
-        return radius.multiply(new BigDecimal(2*Math.PI));
+    static double areaOfCircle(BigDecimal radius) {
+        return radius.multiply(new BigDecimal(2*Math.PI)).doubleValue();
         //throw new NoCodeWrittenException();
     }
 
-    static BigDecimal areaOfRectangle(BigDecimal length, BigDecimal width) {
-        return length.multiply(width);
+    static double areaOfRectangle(BigDecimal length, BigDecimal width) {
+        return length.multiply(width).doubleValue();
         //throw new NoCodeWrittenException();
     }
 
-    static BigDecimal areaOfTriangle(BigDecimal base, BigDecimal h) {
-        return base.multiply(h).multiply(new BigDecimal("0.5"));
+    static double areaOfTriangle(BigDecimal base, BigDecimal h) {
+        return base.multiply(h).multiply(new BigDecimal("0.5")).doubleValue();
         //throw new NoCodeWrittenException();
     }
 }
